backend refactor, datacenter filtering, model crud validation (#211)
* removed unnecesary validation method from views

* updated rack multiple creation/deletion endpoint to support datacenters

* finished datacenter filtering, going to refactor now

* refactored asset management

* reformate code

* added skeleton for network graph serializers, fixed combined sort for assets by racknum, endpoint for available racknums in a given datacenter

* added stub method to get network connection graph on asset detail
