+++ ass_man/import_manager.py
@@ -2,6 +2,7 @@ import csv, io
 from ass_man.models import Model, Asset, Rack
 from rest_framework.response import Response
 
+
 def import_model_file(request):
     file = request.FILES['file']
     reader = csv.DictReader(io.StringIO(file.read().decode('utf-8-sig')))
@@ -23,16 +24,16 @@ def import_model_file(request):
         if disp_col.startswith('#'):
             disp_col = disp_col[1:]
         if model is None:
-            new_model = Model(vendor=row['vendor'], model_number=row['model_number'],height=row['height'], \
-            cpu=row['cpu'], storage=row['storage'], comment=row['comment'])
+            new_model = Model(vendor=row['vendor'], model_number=row['model_number'], height=row['height'], \
+                              cpu=row['cpu'], storage=row['storage'], comment=row['comment'])
             if disp_col:
-                new_model.display_color=disp_col
+                new_model.display_color = disp_col
             if row['network_ports']:
-                new_model.network_ports=row['network_ports']
+                new_model.network_ports = row['network_ports']
             if row['power_ports']:
-                new_model.power_ports=row['power_ports']
+                new_model.power_ports = row['power_ports']
             if row['memory']:
-                new_model.memory=row['memory']
+                new_model.memory = row['memory']
             models_to_create.append(new_model)
             continue
         if str(model.height) != row['height']:
@@ -102,43 +103,44 @@ def import_model_file(request):
         if should_update:
             models_to_update.append(model)
         if override:
-            overriden+=1
+            overriden += 1
         else:
-            ignored+=1
+            ignored += 1
             models_to_ignore.append(model)
 
     if overriden > 0 and not should_override:
-        err_message = "Do you want to overwrite the following "\
-        "fields: "
+        err_message = "Do you want to overwrite the following " \
+                      "fields: "
         count = 0
         for field in fields_overriden.keys():
             err_message += "For " + field + " overwrite " + str(fields_overriden[field][0]) \
-            + " with " + fields_overriden[field][1] + ". "
+                           + " with " + fields_overriden[field][1] + ". "
 
         return Response({
-            'Warning' : err_message,
+            'Warning': err_message,
         }, status=status.HTTP_400_BAD_REQUEST)
 
     created_models = ''
     for model in models_to_create:
         model.save()
-        created_models += model.vendor+' '+model.model_number+' ,'
+        created_models += model.vendor + ' ' + model.model_number + ' ,'
     updated_models = ''
     for model in models_to_update:
         model.save()
-        updated_models += model.vendor+' '+model.model_number+' ,'
+        updated_models += model.vendor + ' ' + model.model_number + ' ,'
     ignored_models = ''
     for model in models_to_ignore:
-        ignored_models += model.vendor+' '+model.model_number+' ,'
+        ignored_models += model.vendor + ' ' + model.model_number + ' ,'
 
     return Response({
-    'Number of models created': (len(models_to_create)),
-    'Number of models ignored': ignored,
-    'Number of models updated': overriden,
-    'Models created': created_models,
-    'Models updated': updated_models,
-    'Models ignored': ignored_models
+        'Number of models created': (len(models_to_create)),
+        'Number of models ignored': ignored,
+        'Number of models updated': overriden,
+        'Models created': created_models,
+        'Models updated': updated_models,
+        'Models ignored': ignored_models
     })
+
 
 def import_asset_file(request):
     file = request.FILES['file']
@@ -189,17 +191,17 @@ def import_asset_file(request):
                     uncreated_objects['user'].append(row['owner'])
                     dont_add = True
                 else:
-                    owner=None
+                    owner = None
             if not dont_add:
-                asset = Asset(model=model, hostname=row['hostname'],\
-                rack=rack, rack_u=row['rack_position'], owner=owner, comment=row['comment'])
-                for i in range(int(row['rack_position']), int(row['rack_position'])+asset.model.height):
+                asset = Asset(model=model, hostname=row['hostname'], \
+                              rack=rack, rack_u=row['rack_position'], owner=owner, comment=row['comment'])
+                for i in range(int(row['rack_position']), int(row['rack_position']) + asset.model.height):
                     curr_asset = getattr(rack, 'u{}'.format(i))
                     if curr_asset is not None:
                         blocked = True
-                        blocked_assets[asset.hostname] =row['rack']+"_u"+row['rack_position']
+                        blocked_assets[asset.hostname] = row['rack'] + "_u" + row['rack_position']
 
-                for i in range(int(row['rack_position']), int(row['rack_position'])+asset.model.height):
+                for i in range(int(row['rack_position']), int(row['rack_position']) + asset.model.height):
                     setattr(rack, 'u{}'.format(i), asset)
                 racks_to_save.append(rack)
 
@@ -235,16 +237,16 @@ def import_asset_file(request):
 
             if should_override:
                 blocked = False
-                for i in range(int(row['rack_position']), int(row['rack_position'])+asset.model.height):
+                for i in range(int(row['rack_position']), int(row['rack_position']) + asset.model.height):
                     curr_asset = getattr(rack, 'u{}'.format(i))
                     if curr_asset is not None:
                         blocked = True
-                        blocked_assets[asset.hostname] =row['rack']+"_u"+row['rack_position']
+                        blocked_assets[asset.hostname] = row['rack'] + "_u" + row['rack_position']
                 if not blocked:
                     asset.rack = rack
-                    for i in range(old_u, old_u+asset.model.height):
+                    for i in range(old_u, old_u + asset.model.height):
                         setattr(rack, 'u{}'.format(i), None)
-                    for i in range(int(row['rack_position']), int(row['rack_position'])+asset.model.height):
+                    for i in range(int(row['rack_position']), int(row['rack_position']) + asset.model.height):
                         setattr(rack, 'u{}'.format(i), asset)
                     racks_to_save.append(rack)
                 should_update = True
@@ -268,19 +270,19 @@ def import_asset_file(request):
                 except Rack.DoesNotExist:
                     rack = None
                 blocked = False
-                for i in range(int(row['rack_position']), int(row['rack_position'])+asset.model.height):
+                for i in range(int(row['rack_position']), int(row['rack_position']) + asset.model.height):
                     curr_asset = getattr(rack, 'u{}'.format(row['rack_position']))
                     if curr_asset is not None:
                         blocked = True
-                        blocked_assets[asset.hostname] =row['rack']+"_u"+row['rack_position']
+                        blocked_assets[asset.hostname] = row['rack'] + "_u" + row['rack_position']
                 if not blocked:
                     old_u = asset.rack_u
                     asset.rack_u = row['rack_position']
                     print('old u ' + str(old_u))
                     print('new u ' + str(asset.rack_u))
-                    for i in range(old_u, old_u+asset.model.height):
+                    for i in range(old_u, old_u + asset.model.height):
                         setattr(rack, 'u{}'.format(i), None)
-                    for i in range(int(row['rack_position']), int(row['rack_position'])+asset.model.height):
+                    for i in range(int(row['rack_position']), int(row['rack_position']) + asset.model.height):
                         setattr(rack, 'u{}'.format(i), asset)
                     racks_to_save.append(rack)
                 should_update = True
@@ -320,59 +322,59 @@ def import_asset_file(request):
         if should_update:
             assets_to_update.append(asset)
         if override:
-            overriden+=1
+            overriden += 1
         else:
-            ignored+=1
+            ignored += 1
             assets_to_ignore.append(asset)
 
     if len(uncreated_objects['model']) > 0 or len(uncreated_objects['rack']) > 0 or len(uncreated_objects['user']) > 0:
         err_message = "The following objects were referenced, but have not been created. "
         for i in uncreated_objects.keys():
-            err_message+= i + ": "
+            err_message += i + ": "
             for j in uncreated_objects[i]:
-                err_message+= j + ", "
-            err_message+=". "
+                err_message += j + ", "
+            err_message += ". "
         return Response({
-            'Warning' : err_message,
+            'Warning': err_message,
         }, status=status.HTTP_400_BAD_REQUEST)
 
     err_message = "The following assets are blocked for placement: "
     if len(blocked_assets.keys()) > 0:
         for inst in blocked_assets.keys():
-            err_message+=inst + " at " + blocked_assets[inst] + ". "
+            err_message += inst + " at " + blocked_assets[inst] + ". "
         return Response({
-            'Warning' : err_message,
+            'Warning': err_message,
         }, status=status.HTTP_400_BAD_REQUEST)
 
     if overriden > 0 and not should_override:
-        err_message = "Do you want to overwrite the following "\
-        "fields: "
+        err_message = "Do you want to overwrite the following " \
+                      "fields: "
         count = 0
         for field in fields_overriden.keys():
             err_message += "For " + field + " overwrite " + str(fields_overriden[field][0]) \
-            + " with " + fields_overriden[field][1] + ". "
+                           + " with " + fields_overriden[field][1] + ". "
 
         return Response({
-            'Warning' : err_message,
+            'Warning': err_message,
         }, status=status.HTTP_400_BAD_REQUEST)
     created_assets = ''
     updated_assets = ''
     ignored_assets = ''
     for asset in assets_to_create:
         asset.save()
-        created_assets+=asset.hostname + ", "
+        created_assets += asset.hostname + ", "
     for asset in assets_to_update:
         asset.save()
-        updated_assets+=asset.hostname + ", "
+        updated_assets += asset.hostname + ", "
     for asset in assets_to_ignore:
-        ignored_assets+=asset.hostname + ", "
+        ignored_assets += asset.hostname + ", "
     for rack in racks_to_save:
         rack.save()
     return Response({
-    'Number of assets created': (len(assets_to_create)),
-    'Number of assets ignored': ignored,
-    'Number of assets updated': overriden,
-    'Created assets':created_assets,
-    'Updated assets':updated_assets,
-    'Ignored assets':ignored_assets
+        'Number of assets created': (len(assets_to_create)),
+        'Number of assets ignored': ignored,
+        'Number of assets updated': overriden,
+        'Created assets': created_assets,
+        'Updated assets': updated_assets,
+        'Ignored assets': ignored_assets
     })
