+++ ass_man/tests/tests.py
@@ -3,5 +3,3 @@ from django.test import TestCase
 # Create your tests here.
 # Reference: https://docs.djals
 # ª ngoproject.com/en/3.0/topics/testing/overview/
-
-
