+++ requirements.txt
@@ -1,12 +1,20 @@
 asgiref==3.2.3
+astroid==2.3.3
 dj-database-url==0.5.0
 Django==3.0.2
 django-cors-headers==3.2.1
 django-filter==2.2.0
 djangorestframework==3.11.0
 gunicorn==20.0.4
+isort==4.3.21
+lazy-object-proxy==1.4.3
+mccabe==0.6.1
 psycopg2-binary==2.8.4
+PyJWT==1.7.1
+pylint==2.4.4
 python-dotenv==0.10.5
 pytz==2019.3
+six==1.14.0
 sqlparse==0.3.0
 whitenoise==5.0.1
+wrapt==1.11.2
